+++ src/cdi_analytic/Pseudo_Line_Base.cc
@@ -67,7 +67,8 @@ void Pseudo_Line_Base::setup_(double emin, double emax) {
     if (nu0_ < 0) {
       size_t N = continuum_table_.size();
       if (N > 0) {
-        C = continuum_table_[static_cast<unsigned int>(edge_[i] * N / emax)];
+        C = continuum_table_[static_cast<unsigned int>(
+            edge_[i] * static_cast<double>(N) / emax)];
       } else {
         C = (*continuum_)(vector<double>(1, edge_[i]));
       }
@@ -225,7 +226,8 @@ double Pseudo_Line_Base::monoOpacity(double const x, double const T) const {
   if (nu0_ < 0) {
     size_t N = continuum_table_.size();
     if (N > 0) {
-      Result = continuum_table_[static_cast<unsigned int>(x * N / emax_)];
+      Result = continuum_table_[static_cast<unsigned int>(
+          x * static_cast<double>(N) / emax_)];
     } else {
       Result = (*continuum_)(vector<double>(1, x));
     }
